New package: fa-icons
The Awesome Icon library is a free, high-quality icon library which
defines in version 6.1.1 more than 2,000 free icons. It can be
included in all kind of packages, and it does not depend on a CSS
toolkit. It defines an icon font and SVG icons as well. The
https://fontawesome.com website lists in addition to the Free version
also a "Pro" and a "Pro Max" including more than 13K icons.

The package supports installation from CDN and a local installation,
switchable via the site-wide administration link for the package in
/acs-admin/package/fa-icons/.

The package is also already integrated with the <adp:icon> markup
to support generic pages using icon sets depending on the "IconSet"
definition (switchable per subsite).